api/requirements.txt
# Core Dependencies
streamlit
fastapi
uvicorn
python-multipart
numpy
pandas

# ML/AI Dependencies
sentence-transformers
faiss-cpu
scikit-learn
google-generativeai
openai

# Document Processing
pymupdf

# Environment Variables
python-dotenv
